Question1/pyproject.toml
[tool.poetry]
name = "question1"
version = "0.1.0"
description = ""
authors = ["SadineniAbhi <abhisadineni@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.36.0"
openai = "^1.35.3"
google-generativeai = "^0.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
